Move add Update entities to UniFi controller
